comment cleanup, made TypeAssignment public
