+++ src/Bicep.Core/TypeSystem/ParamsTypeManager.cs
@@ -11,16 +11,17 @@ namespace Bicep.Core.TypeSystem
     public class ParamsTypeManager : IParamsTypeManager
     {
         private readonly ParamsTypeAssignmentVisitor paramsTypeAssignmentVisitor;
+
         public ParamsTypeManager(ParamBinder binder)
         {
             this.paramsTypeAssignmentVisitor = new ParamsTypeAssignmentVisitor(this, binder);
         }
-
-        // should return the corresponding declared type of parameter "foo" in the Bicep file
+        
         public TypeSymbol GetDeclaredType(SyntaxBase syntax)
         {
             // use ParameterAssignmentSyntax to find Bicep SemanticModel, call GetType() to get ParameterDeclarationSyntax
             throw new NotImplementedException();
+            
         }
 
         public TypeSymbol GetTypeInfo(SyntaxBase syntax)
